Optimus, hidden unnecessary method of sdk
