+++ native_contract.go
@@ -1,11 +1,14 @@
 package ontology_go_sdk
 
 import (
+	sdkcom "github.com/ontio/ontology-go-sdk/common"
 	"github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology/account"
 	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/core/types"
+	httpcom "github.com/ontio/ontology/http/base/common"
 	"github.com/ontio/ontology/smartcontract/service/native/ont"
+	"fmt"
 )
 
 var (
@@ -19,19 +22,76 @@ var (
 )
 
 type NativeContract struct {
-	Ont *Ont
-	Ong *Ong
+	ontSdk *OntologySdk
+	Ont    *Ont
+	Ong    *Ong
 }
 
 func newNativeContract(ontSdk *OntologySdk) *NativeContract {
-	return &NativeContract{
-		Ont: &Ont{ontSdk: ontSdk},
-		Ong: &Ong{ontSdk: ontSdk},
+	native := &NativeContract{ontSdk: ontSdk}
+	native.Ont = &Ont{native: native, ontSdk: ontSdk}
+	native.Ong = &Ong{native: native, ontSdk: ontSdk}
+	return native
+}
+
+func (this *NativeContract) NewNativeInvokeTransaction(
+	gasPrice,
+	gasLimit uint64,
+	version byte,
+	contractAddress common.Address,
+	method string,
+	params []interface{},
+) (*types.Transaction, error) {
+	if params == nil {
+		params = make([]interface{}, 0, 1)
 	}
+	//Params cannot empty, if params is empty, fulfil with empty string
+	if len(params) == 0 {
+		params = append(params, "")
+	}
+	invokeCode, err := httpcom.BuildNativeInvokeCode(contractAddress, version, method, params)
+	if err != nil {
+		return nil, fmt.Errorf("BuildNativeInvokeCode error:%s", err)
+	}
+	return sdkcom.NewInvokeTransaction(gasPrice, gasLimit, invokeCode), nil
+}
+
+func (this *NativeContract) InvokeNativeContract(
+	gasPrice,
+	gasLimit uint64,
+	singer *account.Account,
+	version byte,
+	contractAddress common.Address,
+	method string,
+	params []interface{},
+) (common.Uint256, error) {
+	tx, err := this.NewNativeInvokeTransaction(gasPrice, gasLimit, version, contractAddress, method, params)
+	if err != nil {
+		return common.UINT256_EMPTY, err
+	}
+	err = this.ontSdk.SignToTransaction(tx, singer)
+	if err != nil {
+		return common.UINT256_EMPTY, err
+	}
+	return this.ontSdk.SendTransaction(tx)
+}
+
+func (this *NativeContract) PreExecInvokeNativeContract(
+	contractAddress common.Address,
+	version byte,
+	method string,
+	params []interface{},
+) (*sdkcom.PreExecResult, error) {
+	tx, err := this.NewNativeInvokeTransaction(0, 0, version, contractAddress, method, params)
+	if err != nil {
+		return nil, err
+	}
+	return this.ontSdk.PreExecTransaction(tx)
 }
 
 type Ont struct {
 	ontSdk *OntologySdk
+	native *NativeContract
 }
 
 func (this *Ont) NewTransferTransaction(gasPrice, gasLimit uint64, from, to common.Address, amount uint64) (*types.Transaction, error) {
@@ -56,7 +116,7 @@ func (this *Ont) Transfer(gasPrice, gasLimit uint64, from *account.Account, to c
 }
 
 func (this *Ont) NewMultiTransferTransaction(gasPrice, gasLimit uint64, states []*ont.State) (*types.Transaction, error) {
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONT_CONTRACT_VERSION,
@@ -84,7 +144,7 @@ func (this *Ont) NewTransferFromTransaction(gasPrice, gasLimit uint64, sender, f
 		To:     to,
 		Value:  amount,
 	}
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONT_CONTRACT_VERSION,
@@ -112,7 +172,7 @@ func (this *Ont) NewApproveTransaction(gasPrice, gasLimit uint64, from, to commo
 		To:    to,
 		Value: amount,
 	}
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONT_CONTRACT_VERSION,
@@ -139,7 +199,7 @@ func (this *Ont) Allowance(from, to common.Address) (uint64, error) {
 		From common.Address
 		To   common.Address
 	}
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.ALLOWANCE_NAME,
@@ -156,7 +216,7 @@ func (this *Ont) Allowance(from, to common.Address) (uint64, error) {
 }
 
 func (this *Ont) Symbol() (string, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.SYMBOL_NAME,
@@ -169,7 +229,7 @@ func (this *Ont) Symbol() (string, error) {
 }
 
 func (this *Ont) BalanceOf(address common.Address) (uint64, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.BALANCEOF_NAME,
@@ -186,7 +246,7 @@ func (this *Ont) BalanceOf(address common.Address) (uint64, error) {
 }
 
 func (this *Ont) Name() (string, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.NAME_NAME,
@@ -199,7 +259,7 @@ func (this *Ont) Name() (string, error) {
 }
 
 func (this *Ont) Decimals() (byte, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.DECIMALS_NAME,
@@ -216,7 +276,7 @@ func (this *Ont) Decimals() (byte, error) {
 }
 
 func (this *Ont) TotalSupply() (uint64, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONT_CONTRACT_ADDRESS,
 		ONT_CONTRACT_VERSION,
 		ont.TOTAL_SUPPLY_NAME,
@@ -234,6 +294,7 @@ func (this *Ont) TotalSupply() (uint64, error) {
 
 type Ong struct {
 	ontSdk *OntologySdk
+	native *NativeContract
 }
 
 func (this *Ong) NewTransferTransaction(gasPrice, gasLimit uint64, from, to common.Address, amount uint64) (*types.Transaction, error) {
@@ -258,7 +319,7 @@ func (this *Ong) Transfer(gasPrice, gasLimit uint64, from *account.Account, to c
 }
 
 func (this *Ong) NewMultiTransferTransaction(gasPrice, gasLimit uint64, states []*ont.State) (*types.Transaction, error) {
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONG_CONTRACT_VERSION,
@@ -286,7 +347,7 @@ func (this *Ong) NewTransferFromTransaction(gasPrice, gasLimit uint64, sender, f
 		To:     to,
 		Value:  amount,
 	}
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONG_CONTRACT_VERSION,
@@ -314,7 +375,7 @@ func (this *Ong) NewApproveTransaction(gasPrice, gasLimit uint64, from, to commo
 		To:    to,
 		Value: amount,
 	}
-	return this.ontSdk.NewNativeInvokeTransaction(
+	return this.native.NewNativeInvokeTransaction(
 		gasPrice,
 		gasLimit,
 		ONG_CONTRACT_VERSION,
@@ -341,7 +402,7 @@ func (this *Ong) Allowance(from, to common.Address) (uint64, error) {
 		From common.Address
 		To   common.Address
 	}
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.ALLOWANCE_NAME,
@@ -358,7 +419,7 @@ func (this *Ong) Allowance(from, to common.Address) (uint64, error) {
 }
 
 func (this *Ong) Symbol() (string, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.SYMBOL_NAME,
@@ -371,7 +432,7 @@ func (this *Ong) Symbol() (string, error) {
 }
 
 func (this *Ong) BalanceOf(address common.Address) (uint64, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.BALANCEOF_NAME,
@@ -388,7 +449,7 @@ func (this *Ong) BalanceOf(address common.Address) (uint64, error) {
 }
 
 func (this *Ong) Name() (string, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.NAME_NAME,
@@ -401,7 +462,7 @@ func (this *Ong) Name() (string, error) {
 }
 
 func (this *Ong) Decimals() (byte, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.DECIMALS_NAME,
@@ -418,7 +479,7 @@ func (this *Ong) Decimals() (byte, error) {
 }
 
 func (this *Ong) TotalSupply() (uint64, error) {
-	preResult, err := this.ontSdk.PreExecInvokeNativeContract(
+	preResult, err := this.native.PreExecInvokeNativeContract(
 		ONG_CONTRACT_ADDRESS,
 		ONG_CONTRACT_VERSION,
 		ont.TOTAL_SUPPLY_NAME,
@@ -433,4 +494,3 @@ func (this *Ong) TotalSupply() (uint64, error) {
 	}
 	return balance.Uint64(), nil
 }
-
