+++ neovm_contract.go
@@ -1,0 +1,122 @@
+package ontology_go_sdk
+
+import (
+	"encoding/hex"
+	"fmt"
+	sdkcom "github.com/ontio/ontology-go-sdk/common"
+	httpcom "github.com/ontio/ontology/http/base/common"
+	"github.com/ontio/ontology/account"
+	"github.com/ontio/ontology/common"
+	"github.com/ontio/ontology/core/payload"
+	"github.com/ontio/ontology/core/types"
+	"time"
+)
+
+type NeoVMContract struct {
+	ontSdk *OntologySdk
+}
+
+func newNeoVMContract(ontSdk *OntologySdk) *NeoVMContract {
+	return &NeoVMContract{
+		ontSdk: ontSdk,
+	}
+}
+
+func (this *NeoVMContract) NewDeployNeoVMCodeTransaction(gasPrice, gasLimit uint64, contract *sdkcom.SmartContract) *types.Transaction {
+	deployPayload := &payload.DeployCode{
+		Code:        contract.Code,
+		NeedStorage: contract.NeedStorage,
+		Name:        contract.Name,
+		Version:     contract.Version,
+		Author:      contract.Author,
+		Email:       contract.Email,
+		Description: contract.Description,
+	}
+	tx := &types.Transaction{
+		Version:  sdkcom.VERSION_TRANSACTION,
+		TxType:   types.Deploy,
+		Nonce:    uint32(time.Now().Unix()),
+		Payload:  deployPayload,
+		GasPrice: gasPrice,
+		GasLimit: gasLimit,
+		Sigs:     make([]*types.Sig, 0, 0),
+	}
+	return tx
+}
+
+//DeploySmartContract Deploy smart contract to ontology
+func (this *NeoVMContract) DeployNeoVMSmartContract(
+	gasPrice,
+	gasLimit uint64,
+	singer *account.Account,
+	needStorage bool,
+	code,
+	name,
+	version,
+	author,
+	email,
+	desc string) (common.Uint256, error) {
+
+	invokeCode, err := hex.DecodeString(code)
+	if err != nil {
+		return common.UINT256_EMPTY, fmt.Errorf("code hex decode error:%s", err)
+	}
+	tx := this.NewDeployNeoVMCodeTransaction(gasPrice, gasLimit, &sdkcom.SmartContract{
+		Code:        invokeCode,
+		NeedStorage: needStorage,
+		Name:        name,
+		Version:     version,
+		Author:      author,
+		Email:       email,
+		Description: desc,
+	})
+	err = this.ontSdk.SignToTransaction(tx, singer)
+	if err != nil {
+		return common.Uint256{}, err
+	}
+	txHash, err := this.ontSdk.SendTransaction(tx)
+	if err != nil {
+		return common.Uint256{}, fmt.Errorf("SendRawTransaction error:%s", err)
+	}
+	return txHash, nil
+}
+
+func (this *NeoVMContract) NewNeoVMInvokeTransaction(
+	gasPrice,
+	gasLimit uint64,
+	contractAddress common.Address,
+	params []interface{},
+) (*types.Transaction, error) {
+	invokeCode, err := httpcom.BuildNeoVMInvokeCode(contractAddress, params)
+	if err != nil {
+		return nil, err
+	}
+	return sdkcom.NewInvokeTransaction(gasPrice, gasLimit, invokeCode), nil
+}
+
+func (this *NeoVMContract) InvokeNeoVMContract(
+	gasPrice,
+	gasLimit uint64,
+	signer *account.Account,
+	contractAddress common.Address,
+	params []interface{}) (common.Uint256, error) {
+	tx, err := this.NewNeoVMInvokeTransaction(gasPrice, gasLimit, contractAddress, params)
+	if err != nil {
+		return common.UINT256_EMPTY, fmt.Errorf("NewNeoVMInvokeTransaction error:%s", err)
+	}
+	err = this.ontSdk.SignToTransaction(tx, signer)
+	if err != nil {
+		return common.UINT256_EMPTY, err
+	}
+	return this.ontSdk.SendTransaction(tx)
+}
+
+func (this *NeoVMContract) PreExecInvokeNeoVMContract(
+	contractAddress common.Address,
+	params []interface{}) (*sdkcom.PreExecResult, error) {
+	tx, err := this.NewNeoVMInvokeTransaction(0, 0, contractAddress, params)
+	if err != nil {
+		return nil, err
+	}
+	return this.ontSdk.PreExecTransaction(tx)
+}
